+++ core/src/components/SpaceComponent.js
@@ -8,6 +8,15 @@ export class SpaceComponent extends Component {
 		
 		this.child = props.child?.[0] ?? null
 		
+		const all = props.all?.[0]
+		const horizontal = props.horizontal?.[0]
+		const vertical = props.vertical?.[0]
+		
+		this.left = props.left?.[0] ?? horizontal ?? all ?? 0
+		this.right = props.right?.[0] ?? horizontal ?? all ?? 0
+		this.top = props.top?.[0] ?? vertical ?? all ?? 0
+		this.bottom = props.bottom?.[0] ?? vertical ?? all ?? 0
+		
 	}
 	
 	onAfter(a, p){